refactor: remove unnecessary gnb files
