--- apps/web/src/common/modules/navigations/gnb/modules/gnb-menu/modules/dashboard-recent-favorite/type.ts
@@ -1,5 +1,0 @@
-export interface GNBDashboardMenuItem {
-    name: string;
-    dashboardId: string;
-    workspaceId: string;
-}
