--- apps/web/src/common/modules/navigations/gnb/modules/gnb-noti/type.ts
@@ -1,6 +1,0 @@
-export const NOTIFICATION_TYPE_ICONS = {
-    INFO: '',
-    ERROR: 'ic_error-filled',
-    SUCCESS: 'ic_check',
-    WARNING: 'ic_warning-filled',
-} as const;
