--- apps/web/src/common/modules/navigations/gnb/types/type.ts
@@ -1,4 +1,0 @@
-import type { WORKSPACE_LOGO_ICON_THEMES } from '@/common/modules/navigations/gnb/constants/constant';
-
-
-export type GNBLogoIconTheme = typeof WORKSPACE_LOGO_ICON_THEMES[number];
