+++ apps/web/src/common/modules/navigations/top-bar/modules/gnb-search-clone/helper.ts
@@ -1,7 +1,7 @@
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
-import type { SuggestionType } from '@/common/modules/navigations/gnb/modules/gnb-search/config';
+import type { SuggestionType } from '@/common/modules/navigations/top-bar/modules/gnb-search-clone/config';
 
 export const createSearchRecent = async (type: SuggestionType, id: string, workspaceId: string) => {
     try {
