+++ apps/web/src/common/modules/navigations/top-bar/modules/gnb-search-clone/modules/gnb-search-dropdown/modules/GNBSearchServiceTab.vue
@@ -18,14 +18,12 @@ import type { MenuInfo } from '@/lib/menu/config';
 import { MENU_INFO_MAP } from '@/lib/menu/menu-info';
 
 import { useProxyValue } from '@/common/composables/proxy-state';
-import {
-    SUGGESTION_TYPE,
-} from '@/common/modules/navigations/gnb/modules/gnb-search/config';
-import GNBSuggestionList from '@/common/modules/navigations/gnb/modules/GNBSuggestionList.vue';
 import type { SuggestionItem, SuggestionType } from '@/common/modules/navigations/top-bar/modules/gnb-search-clone/config';
+import { SUGGESTION_TYPE } from '@/common/modules/navigations/top-bar/modules/gnb-search-clone/config';
 import { createSearchRecent } from '@/common/modules/navigations/top-bar/modules/gnb-search-clone/helper';
 import { useTopBarSearchStore } from '@/common/modules/navigations/top-bar/modules/gnb-search-clone/store';
 import type { FocusingDirection } from '@/common/modules/navigations/top-bar/modules/gnb-search-clone/type';
+import TopBarSuggestionList from '@/common/modules/navigations/top-bar/modules/TopBarSuggestionList.vue';
 
 
 interface Props {
@@ -141,21 +139,21 @@ watch(() => state.trimmedInputText, debounce(async (trimmedText) => {
          class="g-n-b-search-service-tab"
     >
         <div class="service-item-list">
-            <g-n-b-suggestion-list v-show="!state.inputText.length"
-                                   :items="state.defaultServiceMenuItems || []"
-                                   :input-text="state.inputText"
-                                   :is-focused="state.focusingType === SUGGESTION_TYPE.MENU ? props.isFocused : false"
-                                   :focusing-direction="props.focusingDirection"
-                                   @move-focus-end="handleFocusEnd(SUGGESTION_TYPE.MENU, ...arguments)"
-                                   @select="handleSelect"
+            <top-bar-suggestion-list v-show="!state.inputText.length"
+                                     :items="state.defaultServiceMenuItems || []"
+                                     :input-text="state.inputText"
+                                     :is-focused="state.focusingType === SUGGESTION_TYPE.MENU ? props.isFocused : false"
+                                     :focusing-direction="props.focusingDirection"
+                                     @move-focus-end="handleFocusEnd(SUGGESTION_TYPE.MENU, ...arguments)"
+                                     @select="handleSelect"
             />
-            <g-n-b-suggestion-list v-show="state.serviceMenuList && state.serviceMenuList.length > 0"
-                                   :items="state.serviceMenuList || []"
-                                   :input-text="state.inputText"
-                                   :is-focused="state.focusingType === SUGGESTION_TYPE.MENU ? props.isFocused : false"
-                                   :focusing-direction="props.focusingDirection"
-                                   @move-focus-end="handleFocusEnd(SUGGESTION_TYPE.MENU, ...arguments)"
-                                   @select="handleSelect"
+            <top-bar-suggestion-list v-show="state.serviceMenuList && state.serviceMenuList.length > 0"
+                                     :items="state.serviceMenuList || []"
+                                     :input-text="state.inputText"
+                                     :is-focused="state.focusingType === SUGGESTION_TYPE.MENU ? props.isFocused : false"
+                                     :focusing-direction="props.focusingDirection"
+                                     @move-focus-end="handleFocusEnd(SUGGESTION_TYPE.MENU, ...arguments)"
+                                     @select="handleSelect"
             />
         </div>
     </div>
