+++ apps/web/src/common/modules/navigations/top-bar/modules/gnb-search-clone/type.ts
@@ -1,4 +1,4 @@
-import type { SuggestionItem, SuggestionType } from '@/common/modules/navigations/gnb/modules/gnb-search/config';
+import type { SuggestionItem, SuggestionType } from '@/common/modules/navigations/top-bar/modules/gnb-search-clone/config';
 
 export interface DropdownItem {
     itemType: SuggestionType;
